pyproject.toml
[project]
name = "zumi-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "click>=8.1.8",
    "ipykernel>=6.29.5",
    "numpy>=1.24.4",
    "pyserial>=3.5",
    "python-dotenv>=1.0.1",
    "pyzmq>=27.1.0",
    "requests>=2.32.4",
]
